create EncoderExt trait for multisig-prover
