+++ Cargo.toml
@@ -75,6 +75,7 @@ mockall = "0.12.1"
 msgs-derive = { version = "^1.0.0", path = "packages/msgs-derive" }
 multisig = { version = "^1.2.1", path = "contracts/multisig" }
 multisig-prover = { version = "^1.1.1", path = "contracts/multisig-prover" }
+multisig-prover-api = { version = "0.1.0", path = "packages/multisig-prover-api" }
 num-traits = { version = "0.2.14", default-features = false }
 quote = "1.0.38"
 rand = "0.8.5"
