+++ contracts/multisig-prover/src/encoding/mod.rs
@@ -3,25 +3,34 @@ mod bcs;
 mod stellar_xdr;
 
 use axelar_wasm_std::hash::Hash;
-use cosmwasm_schema::cw_serde;
 use cosmwasm_std::HexBinary;
 use error_stack::Result;
 use multisig::msg::SignerWithSig;
 use multisig::verifier_set::VerifierSet;
+use multisig_prover_api::Encoder;
 
 use crate::error::ContractError;
 use crate::payload::Payload;
 
-#[cw_serde]
-#[derive(Copy)]
-pub enum Encoder {
-    Abi,
-    Bcs,
-    StellarXdr,
+pub trait EncoderExt {
+    fn digest(
+        &self,
+        domain_separator: &Hash,
+        verifier_set: &VerifierSet,
+        payload: &Payload,
+    ) -> Result<Hash, ContractError>;
+
+    fn execute_data(
+        &self,
+        domain_separator: &Hash,
+        verifier_set: &VerifierSet,
+        sigs: Vec<SignerWithSig>,
+        payload: &Payload,
+    ) -> Result<HexBinary, ContractError>;
 }
 
-impl Encoder {
-    pub fn digest(
+impl EncoderExt for Encoder {
+    fn digest(
         &self,
         domain_separator: &Hash,
         verifier_set: &VerifierSet,
@@ -36,7 +45,7 @@ impl Encoder {
         }
     }
 
-    pub fn execute_data(
+    fn execute_data(
         &self,
         domain_separator: &Hash,
         verifier_set: &VerifierSet,
