+++ packages/multisig-prover-api/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "multisig-prover-api"
+version = "0.1.0"
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+axelar-wasm-std = { workspace = true, features = ["derive"] }
+cosmwasm-schema = { workspace = true }
+multisig = { workspace = true, features = ["library"] }
+
+
+[lints]
+workspace = true
