+++ packages/multisig-prover-api/src/encoding.rs
@@ -1,0 +1,9 @@
+use cosmwasm_schema::cw_serde;
+
+#[cw_serde]
+#[derive(Copy)]
+pub enum Encoder {
+    Abi,
+    Bcs,
+    StellarXdr,
+}
