+++ packages/multisig-prover-api/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod encoding;
+pub mod msg;
+
+pub use crate::encoding::Encoder;
